* changed appIcon, Changed code to fix coding style
